(master) WIP

+++ generators/Dash.module.rb
@@ -33,8 +33,10 @@ class Dash
     @docs_root
     @repo
     @docs_dev_branch
+    @icon
+    @queries
 
-    attr_reader :docset_path, :docset_contents_path, :docset_resources_path, :docset_documents_path, :docs_root
+    attr_reader :docset_path, :docset_contents_path, :docset_resources_path, :docset_documents_path, :docs_root, :queries
 
 
     ##
@@ -63,6 +65,15 @@ class Dash
         if options[:display_name].nil?
             options[:display_name] = options[:name]
         end
+        if !options[:icon].nil?
+            if File.exists?(options[:icon])
+                @icon = options[:icon]
+            elsif File.exists?( File.join(ROOT_PATH, options[:icon]) )
+                @icon = File.join(ROOT_PATH, options[:icon])
+            else
+                @icon = nil
+            end
+        end
 
         @name                   = options[:name]
         @display_name           = options[:display_name]
@@ -72,6 +83,7 @@ class Dash
         @docset_resources_path  = File.join(@docset_contents_path, 'Resources')
         @docset_documents_path  = File.join(@docset_resources_path, 'Documents')
         @docs_dev_branch        = 'dev'
+        @queries                = []
 
 
         if File.exists?(@docset_path)
@@ -116,11 +128,22 @@ class Dash
     end
 
     # copies the files in @docs_root into @docset_documents_path.
-    def copy_docs
+    def copy_docs(options = {})
+        if options[:noop]
+            return
+        end
         # make sure that, for some reason, @docs_root != ( '/' | '/one-level' )
         if @docs_root.split(File::SEPARATOR).length > 2 && File.directory?(@docset_documents_path)
+            puts "Copying documentation into docsets..."
+
             FileUtils.rm_r( Dir.glob( File.join(@docset_documents_path, '*') ) )
             FileUtils.cp_r( File.join(@docs_root, '.'), @docset_documents_path )
+
+            # remove .git directory if there was one (which there should be)
+            git_path = File.join(@docset_documents_path, '.git')
+            if File.directory?(git_path)
+                FileUtils.rm_r(git_path, :force => true)
+            end
         else
             puts "(E) Dash.copy_docs: Failed to copy docs to #{@docset_documents_path}."
         end
@@ -172,6 +195,8 @@ class Dash
                 a['id'] = name
             elsif anchor_id != ''
                 a['id'] = anchor_id
+            else
+                a['id'] = ''
             end
 
             return a
@@ -197,11 +222,17 @@ class Dash
     end
 
     # runs query in the docset's sqlite database.
-    def sql_insert(query)
-        if File.directory?(@docset_resources_path)
-            `cd "#{@docset_resources_path}"; sqlite3 docSet.dsidx '#{query}'`
-        else
-            puts "(E) Dash.sql_insert: Not inserting record. Could not find #{@docset_resources_path}."
+    def sql_insert(*args)
+        if args.length == 3
+            sql_insert(get_sql_insert(*args))
+
+        elsif args.length == 1
+            @queries.push(args[0])
+            # if File.directory?(@docset_resources_path)
+            #     `cd "#{@docset_resources_path}"; sqlite3 docSet.dsidx '#{args[0]}'`
+            # else
+            #     puts "(E) Dash.sql_insert: Not inserting record. Could not find #{@docset_resources_path}."
+            # end
         end
     end
 
@@ -218,7 +249,68 @@ class Dash
         end
     end
 
-    # copies the plist template into the appripriate docset location
+    # execute sql statements in the +@queries+ array
+    def sql_execute(options = {})
+        if @queries.length > 0
+            do_queries = true
+            queries = @queries.clone
+
+            if options.length
+                if options[:noop] == true
+                    do_queries = false
+                    queries.slice!(0..1)
+                    puts "Total of #{queries.length} queries."
+                    if options[:filter].is_a?(Hash)
+                        filters = options[:filter]
+                        if filters[:type]
+                            puts "Filtering by type: #{filters[:type]}"
+                            patt = Regexp.new(', "' + filters[:type] + '",', Regexp::IGNORECASE)
+                            queries.length && queries.map! {|query| query.match(patt) ? query : nil }
+                        end
+                        queries.compact!
+                        if filters[:name]
+                            puts "Filtering by name: #{filters[:name]}"
+                            patt = Regexp.new('\("[^"]*' + filters[:name] + '[^"]*",', Regexp::IGNORECASE | Regexp::EXTENDED)
+                            queries.length && queries.map! {|query| query.match(patt) ? query : nil }
+                        end
+                        queries.compact!
+                        if filters[:limit]
+                            puts "Limiting results to: #{filters[:limit]}"
+                            start = 0
+                            finish = start + filters[:limit].to_i - 1
+                            range = start..finish
+                            queries = queries[range]
+                        end
+                        queries.compact!
+                    end
+                    hr = '-' * 50
+                    puts "\n#{hr}"
+                    puts 'Staged SQL statements:'
+                    puts hr
+                    puts queries.join("\n")
+                    puts hr
+                end
+            end
+
+            if do_queries
+                puts "Executing #{queries.length} queries..."
+                if File.directory?(@docset_resources_path)
+                    query_file_path = File.join(RESOURCES_PATH, 'sqlite3.queries')
+                    file = File.new(query_file_path, 'w')
+                    file << @queries.join("\n")
+                    file.close
+
+                    `cd "#{@docset_resources_path}"; sqlite3 docSet.dsidx < '#{query_file_path}'`
+
+                else
+                    puts "(E) Dash.sql_execute: Not executing query. Could not find #{@docset_resources_path}."
+                end
+            end
+
+        end
+    end
+
+    # copies the plist template into the appropriate docset location
     def copy_plist
         if File.directory?(@docset_contents_path)
             plist_src_path = File.join(RESOURCES_PATH, 'Info.plist')
@@ -239,6 +331,20 @@ class Dash
         end
     end
 
+    # copies a user-supplied icon into the appropriate docset location
+    def copy_icon
+        if !@icon.nil?
+            if File.exists?(@icon) && @icon.match(/\.png$/)
+                FileUtils.cp(@icon, File.join(@docset_path, 'icon.png'))
+            else
+                puts "(E) Dash.copy_icon: Bad (not PNG) or missing filename given: #{@icon}"
+            end
+
+        else
+            puts "(E) Dash.copy_icon: No icon to copy."
+        end
+    end
+
     # creates the docset hierarchy, the docset .plist, and initializes the database.
     def create_docset
         if !File.exists?(@docset_path)
@@ -248,8 +354,15 @@ class Dash
             puts "Dash.create_docset: Copying .plist file..."
             copy_plist
 
+            if !@icon.nil?
+                puts "Dash.create_docset: Copying icon image..."
+                copy_icon
+            end
+
             puts "Dash.create_docset: Setting up database..."
             setup_sql
+
+            puts "Dash.create_docset: Docset set-up complete!"
         else
             puts "(W) Dash.create_docset: Docset path already exists. Skipping..."
         end
@@ -288,7 +401,7 @@ class Dash
                     repo = Git::init(@docs_root)
                     # add all files
                     repo.add( :all => true )
-                    repo.commit('(master) initial import')
+                    repo.commit('(master) initial import from Dash generator')
                     repo.branch(@docs_dev_branch).checkout
                 else
                     puts "Not creating docs repo..."
